+++ ms-transactions/src/main/java/com/quandrix/ms_transactions/dto/TransactionResponse.java
@@ -2,7 +2,6 @@ package com.quandrix.ms_transactions.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -13,6 +12,6 @@ public class TransactionResponse {
     private Long buyerId;
     private Long sellerId;
     private String scryfallId;
-    private BigDecimal amount;
+    private Long amount;
     private LocalDateTime completedAt;
 }
